requirements.txt
alpaca-py==0.25.0
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
asyncpg==0.29.0
sqlalchemy==2.0.23
pydantic==2.5.0